removed hardcoded paths
this now relies on the RED system to work. For Clarion8 I have included
a local RED file that has the required paths and which includes the
default RED file for all other normal operations. For C6 you will have
to either edit your RED file to be like my example or put the resource
files somewhere they can be found by C6